+++ tests/gradients/test_parameter_shift_hessian.py
@@ -228,7 +228,538 @@ class TestGenerateOffDiagTapes:
 
 class TestParameterShiftHessian:
     """Test the general functionality of the param_shift_hessian method
-    on the default interface (autograd)"""
+    on the tape level"""
+
+    def test_single_expval(self):
+        """Test that the correct hessian is calculated for a tape with single RX operator
+        and single expectation value output"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array(0.1, requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(0))
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        expected = -np.cos(x)
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, np.ndarray)
+        assert hessian.shape == ()
+        assert np.allclose(expected, hessian)
+
+    def test_single_probs(self):
+        """Test that the correct hessian is calculated for a tape with single RX operator
+        and single probability output"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array(0.1, requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            qml.probs(wires=[0, 1])
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        expected = 0.5 * np.cos(x) * np.array([-1, 0, 0, 1])
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, np.ndarray)
+        assert hessian.shape == (4,)
+        assert np.allclose(expected, hessian)
+
+    def test_multi_expval(self):
+        """Test that the correct hessian is calculated for a tape with single RX operator
+        and multiple expval outputs"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array(0.1, requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(0))
+            qml.expval(qml.Hadamard(1))
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        expected = (-np.cos(x), -np.cos(x) / np.sqrt(2))
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 2
+
+        for hess, exp in zip(hessian, expected):
+            assert isinstance(hess, np.ndarray)
+            assert hess.shape == ()
+            assert np.allclose(hess, exp)
+
+    def test_multi_expval_probs(self):
+        """Test that the correct hessian is calculated for a tape with single RX operator
+        and both expval and probability outputs"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array(0.1, requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(0))
+            qml.probs(wires=[0, 1])
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        expected = (-np.cos(x), 0.5 * np.cos(x) * np.array([-1, 0, 0, 1]))
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 2
+
+        for hess, exp in zip(hessian, expected):
+            assert isinstance(hess, np.ndarray)
+            assert hess.shape == exp.shape
+            assert np.allclose(hess, exp)
+
+    def test_multi_probs(self):
+        """Test that the correct hessian is calculated for a tape with single RX operator
+        and multiple probability outputs"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array(0.1, requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            qml.probs(wires=[0])
+            qml.probs(wires=[0, 1])
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        expected = (0.5 * np.cos(x) * np.array([-1, 1]), 0.5 * np.cos(x) * np.array([-1, 0, 0, 1]))
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 2
+
+        for hess, exp in zip(hessian, expected):
+            assert isinstance(hess, np.ndarray)
+            assert hess.shape == exp.shape
+            assert np.allclose(hess, exp)
+
+    def test_single_expval_multi_params(self):
+        """Test that the correct hessian is calculated for a tape with multiple operators
+        and single expectation value output"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array([0.1, 0.4], requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x[0], wires=0)
+            qml.RY(x[1], wires=1)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(0))
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        expected = ((-np.cos(x[0]), 0), (0, 0))
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 2
+        assert all(isinstance(hess, tuple) for hess in hessian)
+        assert all(len(hess) == 2 for hess in hessian)
+        assert all(
+            all(isinstance(h, np.ndarray) and h.shape == () for h in hess) for hess in hessian
+        )
+
+        assert np.allclose(hessian, expected)
+
+    def test_single_probs_multi_params(self):
+        """Test that the correct hessian is calculated for a tape with multiple operators
+        and single probability output"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array([0.1, 0.4], requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x[0], wires=0)
+            qml.RY(x[1], wires=1)
+            qml.CNOT(wires=[0, 1])
+            qml.probs(wires=[0, 1])
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        a = [
+            np.cos(x[0] / 2) ** 2,
+            np.sin(x[0] / 2) ** 2,
+            np.cos(x[1] / 2) ** 2,
+            np.sin(x[1] / 2) ** 2,
+        ]
+        expected = (
+            (
+                0.5 * np.cos(x[0]) * np.array([-a[2], -a[3], a[3], a[2]]),
+                0.25 * np.sin(x[0]) * np.sin(x[1]) * np.array([1, -1, 1, -1]),
+            ),
+            (
+                0.25 * np.sin(x[0]) * np.sin(x[1]) * np.array([1, -1, 1, -1]),
+                0.5 * np.cos(x[1]) * np.array([-a[0], a[0], a[1], -a[1]]),
+            ),
+        )
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 2
+        assert all(isinstance(hess, tuple) for hess in hessian)
+        assert all(len(hess) == 2 for hess in hessian)
+        assert all(
+            all(isinstance(h, np.ndarray) and h.shape == (4,) for h in hess) for hess in hessian
+        )
+
+        assert np.allclose(hessian, expected)
+
+    def test_multi_expval_multi_params(self):
+        """Test that the correct hessian is calculated for a tape with multiple operators
+        and multiple expval outputs"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array([0.1, 0.4], requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x[0], wires=0)
+            qml.RY(x[1], wires=1)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(0))
+            qml.expval(qml.Hadamard(1))
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        expected = (
+            ((-np.cos(x[0]), 0), (0, 0)),
+            (
+                (
+                    -np.cos(x[0]) * np.cos(x[1]) / np.sqrt(2),
+                    np.sin(x[0]) * np.sin(x[1]) / np.sqrt(2),
+                ),
+                (
+                    np.sin(x[0]) * np.sin(x[1]) / np.sqrt(2),
+                    (-np.sin(x[1]) - np.cos(x[0]) * np.cos(x[1])) / np.sqrt(2),
+                ),
+            ),
+        )
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 2
+
+        for hess, exp in zip(hessian, expected):
+            assert isinstance(hess, tuple)
+            assert len(hess) == 2
+            assert all(isinstance(h, tuple) for h in hess)
+            assert all(len(h) == 2 for h in hess)
+            assert all(all(isinstance(h_, np.ndarray) and h_.shape == () for h_ in h) for h in hess)
+            assert np.allclose(hess, exp)
+
+    def test_multi_expval_probs_multi_params(self):
+        """Test that the correct hessian is calculated for a tape with multiple operators
+        and both expval and probability outputs"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array([0.1, 0.4], requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x[0], wires=0)
+            qml.RY(x[1], wires=1)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(0))
+            qml.probs(wires=[0, 1])
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        a = [
+            np.cos(x[0] / 2) ** 2,
+            np.sin(x[0] / 2) ** 2,
+            np.cos(x[1] / 2) ** 2,
+            np.sin(x[1] / 2) ** 2,
+        ]
+        expected = (
+            ((-np.cos(x[0]), 0), (0, 0)),
+            (
+                (
+                    0.5 * np.cos(x[0]) * np.array([-a[2], -a[3], a[3], a[2]]),
+                    0.25 * np.sin(x[0]) * np.sin(x[1]) * np.array([1, -1, 1, -1]),
+                ),
+                (
+                    0.25 * np.sin(x[0]) * np.sin(x[1]) * np.array([1, -1, 1, -1]),
+                    0.5 * np.cos(x[1]) * np.array([-a[0], a[0], a[1], -a[1]]),
+                ),
+            ),
+        )
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 2
+
+        for hess, exp in zip(hessian, expected):
+            assert isinstance(hess, tuple)
+            assert len(hess) == 2
+            assert all(isinstance(h, tuple) for h in hess)
+            assert all(len(h) == 2 for h in hess)
+            assert all(
+                all(isinstance(h_, np.ndarray) and h_.shape == exp[0][0].shape for h_ in h)
+                for h in hess
+            )
+            assert np.allclose(hess, exp)
+
+    def test_multi_probs_multi_params(self):
+        """Test that the correct hessian is calculated for a tape with multiple operators
+        and multiple probability outputs"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array([0.1, 0.4], requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x[0], wires=0)
+            qml.RY(x[1], wires=1)
+            qml.CNOT(wires=[0, 1])
+            qml.probs(wires=[1])
+            qml.probs(wires=[0, 1])
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        a = [
+            np.cos(x[0] / 2) ** 2,
+            np.sin(x[0] / 2) ** 2,
+            np.cos(x[1] / 2) ** 2,
+            np.sin(x[1] / 2) ** 2,
+        ]
+        expected = (
+            (
+                (
+                    0.5 * np.cos(x[0]) * np.cos(x[1]) * np.array([-1, 1]),
+                    0.5 * np.sin(x[0]) * np.sin(x[1]) * np.array([1, -1]),
+                ),
+                (
+                    0.5 * np.sin(x[0]) * np.sin(x[1]) * np.array([1, -1]),
+                    0.5 * np.cos(x[0]) * np.cos(x[1]) * np.array([-1, 1]),
+                ),
+            ),
+            (
+                (
+                    0.5 * np.cos(x[0]) * np.array([-a[2], -a[3], a[3], a[2]]),
+                    0.25 * np.sin(x[0]) * np.sin(x[1]) * np.array([1, -1, 1, -1]),
+                ),
+                (
+                    0.25 * np.sin(x[0]) * np.sin(x[1]) * np.array([1, -1, 1, -1]),
+                    0.5 * np.cos(x[1]) * np.array([-a[0], a[0], a[1], -a[1]]),
+                ),
+            ),
+        )
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 2
+
+        for hess, exp in zip(hessian, expected):
+            assert isinstance(hess, tuple)
+            assert len(hess) == 2
+            assert all(isinstance(h, tuple) for h in hess)
+            assert all(len(h) == 2 for h in hess)
+            assert all(
+                all(isinstance(h_, np.ndarray) and h_.shape == exp[0][0].shape for h_ in h)
+                for h in hess
+            )
+            assert np.allclose(hess, exp)
+
+    def test_multi_params_argnum(self):
+        """Test that the correct hessian is calculated for a tape with multiple operators
+        but not all parameters trainable"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array([0.1, 0.4, 0.7], requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x[0], wires=0)
+            qml.RY(x[1], wires=1)
+            qml.RY(x[2], wires=0)
+            qml.CNOT(wires=[0, 1])
+            qml.expval(qml.PauliZ(0))
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        expected = ((0, 0, 0), (0, 0, 0), (0, 0, -np.cos(x[2] + x[0])))
+
+        tapes, fn = qml.gradients.param_shift_hessian(tape, argnum=(1, 2))
+        hessian = fn(qml.execute(tapes, dev, gradient_fn=None))
+
+        assert isinstance(hessian, tuple)
+        assert len(hessian) == 3
+        assert all(isinstance(hess, tuple) for hess in hessian)
+        assert all(len(hess) == 3 for hess in hessian)
+        assert all(
+            all(isinstance(h, np.ndarray) and h.shape == () for h in hess) for hess in hessian
+        )
+
+        assert np.allclose(hessian, expected)
+
+    def test_state_error(self):
+        """Test that an error is raised when computing the gradient of a tape
+        that returns state"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array(0.1, requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            qml.state()
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        msg = "Computing the Hessian of circuits that return the state is not supported"
+        with pytest.raises(ValueError, match=msg):
+            tapes, fn = qml.gradients.param_shift_hessian(tape)
+
+    def test_variance_error(self):
+        """Test that an error is raised when computing the gradient of a tape
+        that returns variance"""
+        dev = qml.device("default.qubit", wires=2)
+
+        x = np.array(0.1, requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RY(x, wires=0)
+            qml.CNOT(wires=[0, 1])
+            qml.var(qml.PauliZ(0))
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        msg = "Computing the Hessian of circuits that return variances is currently not supported"
+        with pytest.raises(ValueError, match=msg):
+            tapes, fn = qml.gradients.param_shift_hessian(tape)
+
+    @pytest.mark.parametrize("num_measurements", [1, 2])
+    def test_no_trainable_params(self, num_measurements):
+        """Test that the correct output and warning is generated in the absence of any trainable
+        parameters"""
+        dev = qml.device("default.qubit", wires=2)
+
+        weights = [0.1, 0.2]
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RX(weights[0], wires=0)
+            qml.RY(weights[1], wires=0)
+            for _ in range(num_measurements):
+                qml.expval(qml.PauliZ(0) @ qml.PauliZ(1))
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        tape.trainable_params = []
+
+        msg = "Attempted to compute the hessian of a tape with no trainable parameters"
+        with pytest.warns(UserWarning, match=msg):
+            tapes, fn = qml.gradients.param_shift_hessian(tape)
+
+        res = fn(qml.execute(tapes, dev, None))
+
+        if num_measurements == 1:
+            res = (res,)
+
+        assert tapes == []
+        assert isinstance(res, tuple)
+        assert len(res) == num_measurements
+        assert all(isinstance(r, np.ndarray) and r.shape == (0,) for r in res)
+
+    @pytest.mark.parametrize("num_measurements", [1, 2])
+    def test_all_zero_grads(self, num_measurements):
+        """Test that the transform works correctly when the diff method for every parameter is
+        identified to be 0, and that no tapes were generated."""
+        dev = qml.device("default.qubit", wires=2)
+
+        class DummyOp(qml.CRZ):
+            grad_method = "0"
+
+        x = np.array(0.1, requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            DummyOp(x, wires=[0, 1])
+            for _ in range(num_measurements):
+                qml.probs(wires=[0, 1])
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        tapes, fn = qml.gradients.param_shift_hessian(tape)
+        res = fn(qml.execute(tapes, dev, None))
+
+        if num_measurements == 1:
+            res = (res,)
+
+        assert tapes == []
+        assert isinstance(res, tuple)
+        assert len(res) == num_measurements
+        assert all(
+            isinstance(r, np.ndarray) and np.allclose(r, np.array([0, 0, 0, 0])) for r in res
+        )
+
+    def test_error_unsupported_op(self):
+        """Test that the correct error is thrown for unsupported operations"""
+
+        dev = qml.device("default.qubit", wires=2)
+
+        class DummyOp(qml.CRZ):
+            grad_method = "F"
+
+        x = np.array([0.1, 0.2, 0.3], requires_grad=True)
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RX(x[0], wires=0)
+            qml.RY(x[1], wires=0)
+            DummyOp(x[2], wires=[0, 1])
+            qml.probs(wires=1)
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        msg = "The parameter-shift Hessian currently does not support the operations"
+        with pytest.raises(ValueError, match=msg):
+            qml.gradients.param_shift_hessian(tape, argnum=[0, 1, 2])(x)
+
+    @pytest.mark.parametrize("argnum", [None, (0,)])
+    def test_error_wrong_diagonal_shifts(self, argnum):
+        """Test that an error is raised if the number of diagonal shifts does
+        not match the required number (`len(trainable_params)` or `len(argnum)`)."""
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RX(0.4, wires=0)
+            qml.CRY(0.9, wires=[0, 1])
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        with pytest.raises(ValueError, match="sets of shift values for diagonal entries"):
+            qml.gradients.param_shift_hessian(tape, argnum=argnum, diagonal_shifts=[])
+
+    @pytest.mark.parametrize("argnum", [None, (0, 1)])
+    def test_error_wrong_offdiagonal_shifts(self, argnum):
+        """Test that an error is raised if the number of offdiagonal shifts does
+        not match the required number (`len(trainable_params)` or `len(argnum)`)."""
+
+        with qml.queuing.AnnotatedQueue() as q:
+            qml.RX(0.4, wires=0)
+            qml.CRY(0.9, wires=[0, 1])
+            qml.RX(-0.4, wires=0)
+
+        tape = qml.tape.QuantumScript.from_queue(q)
+        with pytest.raises(ValueError, match="sets of shift values for off-diagonal entries"):
+            qml.gradients.param_shift_hessian(tape, argnum=argnum, off_diagonal_shifts=[])
+
+
+class TestParameterShiftHessianQNode:
+    """Test the general functionality of the param_shift_hessian method
+    with QNodes on the default interface (autograd)"""
 
     def test_single_two_term_gate(self):
         """Test that the correct hessian is calculated for a QNode with single RX operator
@@ -247,7 +778,6 @@ class TestParameterShiftHessian:
         expected = qml.jacobian(qml.grad(circuit))(x)
         hessian = qml.gradients.param_shift_hessian(circuit)(x)
 
-        assert circuit.interface == "auto"
         assert np.allclose(expected, hessian)
 
     def test_fixed_params(self):
@@ -417,7 +947,7 @@ class TestParameterShiftHessian:
         expected = qml.jacobian(qml.jacobian(circuit))(x)
         hessian = qml.gradients.param_shift_hessian(circuit)(x)
 
-        assert np.allclose(expected, hessian)
+        assert np.allclose(qml.math.transpose(expected, (2, 1, 0)), hessian)
 
     def test_quantum_hessian_shape_vector_input_vector_output(self):
         """Test that the purely "quantum" hessian has the correct shape (1d -> 1d)"""
@@ -434,7 +964,7 @@ class TestParameterShiftHessian:
             return qml.probs(wires=[0, 1])
 
         x = np.array([0.1, 0.2, 0.3], requires_grad=True)
-        shape = (4, 6, 6)  # (num_output_vals, num_gate_args, num_gate_args)
+        shape = (6, 6, 4)  # (num_gate_args, num_gate_args, num_output_vals)
 
         hessian = qml.gradients.param_shift_hessian(circuit, hybrid=False)(x)
 
@@ -459,7 +989,7 @@ class TestParameterShiftHessian:
         expected = qml.jacobian(qml.jacobian(circuit))(x)
         hessian = qml.gradients.param_shift_hessian(circuit)(x)
 
-        assert np.allclose(expected, hessian)
+        assert np.allclose(qml.math.transpose(expected, (2, 1, 0)), hessian)
 
     def test_multiple_two_term_gates_classical_processing(self):
         """Test that the correct hessian is calculated when manipulating parameters (1d -> 1d)"""
@@ -479,7 +1009,7 @@ class TestParameterShiftHessian:
         expected = qml.jacobian(qml.jacobian(circuit))(x)
         hessian = qml.gradients.param_shift_hessian(circuit)(x)
 
-        assert np.allclose(expected, hessian)
+        assert np.allclose(qml.math.transpose(expected, (2, 1, 0)), hessian)
 
     def test_multiple_two_term_gates_matrix_output(self):
         """Test that the correct hessian is calculated for higher dimensional QNode outputs
@@ -487,19 +1017,22 @@ class TestParameterShiftHessian:
 
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="parameter-shift", max_diff=2)
+        @qml.qnode(dev, max_diff=2)
         def circuit(x):
             qml.RX(x[0], wires=0)
             qml.RY(x[1], wires=0)
             qml.CNOT(wires=[0, 1])
             return qml.probs(wires=0), qml.probs(wires=1)
 
+        def cost(x):
+            return qml.math.stack(circuit(x))
+
         x = np.ones([2], requires_grad=True)
 
-        expected = qml.jacobian(qml.jacobian(circuit))(x)
+        expected = qml.jacobian(qml.jacobian(cost))(x)
         hessian = qml.gradients.param_shift_hessian(circuit)(x)
 
-        assert np.allclose(expected, hessian)
+        assert np.allclose(qml.math.transpose(expected, (0, 3, 2, 1)), hessian)
 
     def test_multiple_two_term_gates_matrix_input(self):
         """Test that the correct hessian is calculated for higher dimensional cl. jacobians
@@ -507,7 +1040,7 @@ class TestParameterShiftHessian:
 
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="parameter-shift", max_diff=2)
+        @qml.qnode(dev, max_diff=2)
         def circuit(x):
             qml.RX(x[0, 0], wires=0)
             qml.RY(x[0, 1], wires=0)
@@ -516,18 +1049,21 @@ class TestParameterShiftHessian:
             qml.RY(x[0, 0], wires=0)
             return qml.probs(wires=0), qml.probs(wires=1)
 
+        def cost(x):
+            return qml.math.stack(circuit(x))
+
         x = np.ones([1, 3], requires_grad=True)
 
-        expected = qml.jacobian(qml.jacobian(circuit))(x)
+        expected = qml.jacobian(qml.jacobian(cost))(x)
         hessian = qml.gradients.param_shift_hessian(circuit)(x)
 
-        assert np.allclose(expected, hessian)
+        assert np.allclose(qml.math.transpose(expected, (0, 2, 3, 4, 5, 1)), hessian)
 
     def test_multiple_qnode_arguments_scalar(self):
         """Test that the correct Hessian is calculated with multiple QNode arguments (0D->1D)"""
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="parameter-shift", max_diff=2)
+        @qml.qnode(dev, max_diff=2)
         def circuit(x, y, z):
             qml.RX(x, wires=0)
             qml.RY(y, wires=1)
@@ -554,7 +1090,7 @@ class TestParameterShiftHessian:
         """Test that the correct Hessian is calculated with multiple QNode arguments (1D->1D)"""
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="parameter-shift", max_diff=2)
+        @qml.qnode(dev, max_diff=2)
         def circuit(x, y, z):
             qml.RX(x[0], wires=1)
             qml.RY(y[0], wires=0)
@@ -576,7 +1112,7 @@ class TestParameterShiftHessian:
         circuit.interface = "autograd"
         hessian = qml.gradients.param_shift_hessian(circuit)(x, y, z)
 
-        assert np.allclose(expected, hessian)
+        assert np.allclose(qml.math.transpose(expected, (0, 2, 3, 1)), hessian)
 
     def test_multiple_qnode_arguments_matrix(self):
         """Test that the correct Hessian is calculated with multiple QNode arguments (2D->1D)"""
@@ -604,14 +1140,14 @@ class TestParameterShiftHessian:
         circuit.interface = "autograd"
         hessian = qml.gradients.param_shift_hessian(circuit)(x, y, z)
 
-        assert np.allclose(expected, hessian)
+        assert np.allclose(qml.math.transpose(expected, (0, 2, 3, 4, 5, 1)), hessian)
 
     def test_multiple_qnode_arguments_mixed(self):
         """Test that the correct Hessian is calculated with multiple mixed-shape QNode arguments"""
 
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, max_diff=2, diff_method="parameter-shift")
+        @qml.qnode(dev, max_diff=2)
         def circuit(x, y, z):
             qml.RX(x, wires=0)
             qml.RY(z[0] + z[1], wires=0)
@@ -620,16 +1156,21 @@ class TestParameterShiftHessian:
             qml.CRY(y[0, 1], wires=[0, 1])
             return qml.probs(wires=0), qml.probs(wires=1)
 
+        def cost(x, y, z):
+            return qml.math.stack(circuit(x, y, z))
+
         x = np.array(0.1, requires_grad=True)
         y = np.array([[0.5, 0.6], [0.2, 0.1]], requires_grad=True)
         z = np.array([0.3, 0.4], requires_grad=True)
 
         expected = tuple(
-            qml.jacobian(qml.jacobian(circuit, argnum=i), argnum=i)(x, y, z) for i in range(3)
+            qml.jacobian(qml.jacobian(cost, argnum=i), argnum=i)(x, y, z) for i in range(3)
         )
         hessian = qml.gradients.param_shift_hessian(circuit)(x, y, z)
 
-        assert all(np.allclose(expected[i], hessian[i]) for i in range(3))
+        assert np.allclose(expected[0], hessian[0])
+        assert np.allclose(qml.math.transpose(expected[1], (0, 2, 3, 4, 5, 1)), hessian[1])
+        assert np.allclose(qml.math.transpose(expected[2], (0, 2, 3, 1)), hessian[2])
 
     def test_with_channel(self):
         """Test that the Hessian is correctly computed for circuits
@@ -637,7 +1178,7 @@ class TestParameterShiftHessian:
 
         dev = qml.device("default.mixed", wires=2)
 
-        @qml.qnode(dev, max_diff=2, diff_method="parameter-shift")
+        @qml.qnode(dev, max_diff=2)
         def circuit(x):
             qml.RX(x[1], wires=0)
             qml.RY(x[0], wires=0)
@@ -650,14 +1191,14 @@ class TestParameterShiftHessian:
         expected = qml.jacobian(qml.jacobian(circuit))(x)
         hessian = qml.gradients.param_shift_hessian(circuit)(x)
 
-        assert np.allclose(expected, hessian)
+        assert np.allclose(qml.math.transpose(expected, (2, 1, 0)), hessian)
 
     def test_hessian_transform_is_differentiable(self):
         """Test that the 3rd derivate can be calculated via auto-differentiation (1d -> 1d)"""
 
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, max_diff=3, diff_method="parameter-shift")
+        @qml.qnode(dev, max_diff=3)
         def circuit(x):
             qml.RX(x[1], wires=0)
             qml.RY(x[0], wires=0)
@@ -667,9 +1208,15 @@ class TestParameterShiftHessian:
         x = np.array([0.1, 0.2], requires_grad=True)
 
         expected = qml.jacobian(qml.jacobian(qml.jacobian(circuit)))(x)
-        derivative = qml.jacobian(qml.gradients.param_shift_hessian(circuit))(x)
 
-        assert np.allclose(expected, derivative)
+        def cost_fn(x):
+            hess = qml.gradients.param_shift_hessian(circuit)(x)
+            hess = qml.math.stack([qml.math.stack(row) for row in hess])
+            return hess
+
+        derivative = qml.jacobian(cost_fn)(x)
+
+        assert np.allclose(qml.math.transpose(expected, (1, 2, 0, 3)), derivative)
 
     # Some bounds on the efficiency (device executions) of the hessian for 2-term shift rules:
     # - < jacobian(jacobian())
@@ -677,6 +1224,7 @@ class TestParameterShiftHessian:
     # - <= 3^m                    see arXiv:2008.06517 p. 4
     # here d=2 is the derivative order, m is the number of variational parameters (w.r.t. gate args)
 
+    @pytest.mark.xfail(reason="Update tracker for new return types")
     def test_fewer_device_invocations_scalar_input(self):
         """Test that the hessian invokes less hardware executions than double differentiation
         (0d -> 0d)"""
@@ -702,6 +1250,7 @@ class TestParameterShiftHessian:
         assert hessian_qruns <= 2**2 * 1  # 1 = (1+2-1)C(2)
         assert hessian_qruns <= 3**1
 
+    @pytest.mark.xfail(reason="Update tracker for new return types")
     def test_fewer_device_invocations_vector_input(self):
         """Test that the hessian invokes less hardware executions than double differentiation
         (1d -> 0d)"""
@@ -728,6 +1277,7 @@ class TestParameterShiftHessian:
         assert hessian_qruns <= 2**2 * 3  # 3 = (2+2-1)C(2)
         assert hessian_qruns <= 3**2
 
+    @pytest.mark.xfail(reason="Update tracker for new return types")
     def test_fewer_device_invocations_vector_output(self):
         """Test that the hessian invokes less hardware executions than double differentiation
         (1d -> 1d)"""
@@ -778,30 +1328,6 @@ class TestParameterShiftHessian:
             match=r"The analytic gradient method cannot be used with the parameter\(s\)",
         ):
             qml.gradients.param_shift_hessian(circuit)(x)
-
-    def test_error_unsupported_operation_with_argnum(self):
-        """Test that the correct error is thrown for unsupported operations when
-        argnum is given."""
-
-        dev = qml.device("default.qubit", wires=2)
-
-        class DummyOp(qml.CRZ):
-            grad_method = "F"
-
-        @qml.qnode(dev, max_diff=2, diff_method="parameter-shift")
-        def circuit(x):
-            qml.RX(x[0], wires=0)
-            qml.RY(x[1], wires=0)
-            DummyOp(x[2], wires=[0, 1])
-            return qml.probs(wires=1)
-
-        x = np.array([0.1, 0.2, 0.3], requires_grad=True)
-
-        with pytest.raises(
-            ValueError,
-            match="The parameter-shift Hessian currently does not support",
-        ):
-            qml.gradients.param_shift_hessian(circuit, argnum=[0, 2])(x)
 
     def test_error_unsupported_variance_measurement(self):
         """Test that the correct error is thrown for variance measurements"""
@@ -940,29 +1466,6 @@ class TestParameterShiftHessian:
 
         assert res == ()
 
-    def test_no_trainable_params_tape(self):
-        """Test that the correct ouput and warning is generated in the absence of any trainable
-        parameters"""
-        dev = qml.device("default.qubit", wires=2)
-
-        weights = [0.1, 0.2]
-        with qml.queuing.AnnotatedQueue() as q:
-            qml.RX(weights[0], wires=0)
-            qml.RY(weights[1], wires=0)
-            qml.expval(qml.PauliZ(0) @ qml.PauliZ(1))
-
-        tape = qml.tape.QuantumScript.from_queue(q)
-        # TODO: remove once #2155 is resolved
-        tape.trainable_params = []
-
-        with pytest.warns(UserWarning, match="hessian of a tape with no trainable parameters"):
-            h_tapes, post_processing = qml.gradients.param_shift_hessian(tape)
-        res = post_processing(qml.execute(h_tapes, dev, None))
-
-        assert h_tapes == []
-        assert isinstance(res, np.ndarray)
-        assert res.shape == (1, 0, 0)
-
     def test_all_zero_diff_methods(self):
         """Test that the transform works correctly when the diff method for every parameter is
         identified to be 0, and that no tapes were generated."""
@@ -976,16 +1479,17 @@ class TestParameterShiftHessian:
         params = np.array([0.5, 0.5, 0.5], requires_grad=True)
 
         result = qml.gradients.param_shift_hessian(circuit)(params)
-        assert np.allclose(result, np.zeros((4, 3, 3)), atol=0, rtol=0)
+        assert np.allclose(result, np.zeros((3, 3, 4)), atol=0, rtol=0)
 
         tapes, _ = qml.gradients.param_shift_hessian(circuit.tape)
         assert tapes == []
 
+    @pytest.mark.xfail(reason="Update tracker for new return types")
     def test_f0_argument(self):
         """Test that we can provide the results of a QNode to save on quantum invocations"""
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="parameter-shift", max_diff=2)
+        @qml.qnode(dev, max_diff=2)
         def circuit(x):
             qml.RX(x[0], wires=0)
             qml.RY(x[1], wires=0)
@@ -1037,10 +1541,9 @@ class TestParameterShiftHessian:
         circuits = [qml.QNode(cost, dev) for cost in (cost1, cost2, cost3, cost4, cost5, cost6)]
 
         transform = [qml.math.shape(qml.gradients.param_shift_hessian(c)(x)) for c in circuits]
-        qnode = [qml.math.shape(c(x)) + (3, 3) for c in circuits]
-        expected = [(3, 3), (1, 3, 3), (2, 3, 3), (4, 3, 3), (1, 4, 3, 3), (2, 4, 3, 3)]
+        expected = [(3, 3), (3, 3), (2, 3, 3), (3, 3, 4), (3, 3, 4), (2, 3, 3, 4)]
 
-        assert all(t == q == e for t, q, e in zip(transform, qnode, expected))
+        assert all(t == e for t, e in zip(transform, expected))
 
 
 class TestParamShiftHessianWithKwargs:
@@ -1058,7 +1561,7 @@ class TestParamShiftHessianWithKwargs:
         """Test that diagonal shifts are used and yield the correct Hessian."""
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, max_diff=2, diff_method="parameter-shift")
+        @qml.qnode(dev, max_diff=2)
         def circuit(x):
             qml.RX(x[0], wires=0)
             qml.RY(x[1], wires=0)
@@ -1067,7 +1570,7 @@ class TestParamShiftHessianWithKwargs:
 
         x = np.array([0.6, -0.2], requires_grad=True)
 
-        expected = qml.jacobian(qml.jacobian(circuit))(x)
+        expected = qml.math.transpose(qml.jacobian(qml.jacobian(circuit))(x), (1, 2, 0))
         circuit(x)
         tapes, fn = qml.gradients.param_shift_hessian(
             circuit.qtape, diagonal_shifts=diagonal_shifts
@@ -1102,7 +1605,7 @@ class TestParamShiftHessianWithKwargs:
         """Test that off-diagonal shifts are used and yield the correct Hessian."""
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, max_diff=2, diff_method="parameter-shift")
+        @qml.qnode(dev, max_diff=2)
         def circuit(x):
             qml.RX(x[0], wires=0)
             qml.CRY(x[1], wires=[0, 1])
@@ -1111,7 +1614,7 @@ class TestParamShiftHessianWithKwargs:
 
         x = np.array([0.6, -0.2], requires_grad=True)
 
-        expected = qml.jacobian(qml.jacobian(circuit))(x)
+        expected = qml.math.transpose(qml.jacobian(qml.jacobian(circuit))(x), (1, 2, 0))
         circuit(x)
         tapes, fn = qml.gradients.param_shift_hessian(
             circuit.qtape, off_diagonal_shifts=off_diagonal_shifts
@@ -1181,7 +1684,7 @@ class TestParamShiftHessianWithKwargs:
         """Test that providing an argnum to indicated differentiable parameters works."""
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, max_diff=2, diff_method="parameter-shift")
+        @qml.qnode(dev, max_diff=2)
         def circuit(par):
             qml.RX(par[0], wires=0)
             qml.CRY(par[1], wires=[0, 1])
@@ -1193,7 +1696,9 @@ class TestParamShiftHessianWithKwargs:
 
         expected = qml.jacobian(qml.jacobian(circuit))(par)
         # Set non-argnum argument entries to 0
-        expected = expected * qml.math.array(argnum, dtype=float)[None]
+        expected = qml.math.transpose(
+            expected * qml.math.array(argnum, dtype=float)[None], (1, 2, 0)
+        )
         hessian = qml.gradients.param_shift_hessian(circuit, argnum=argnum)(par)
         assert np.allclose(hessian, expected)
 
@@ -1229,37 +1734,11 @@ class TestParamShiftHessianWithKwargs:
         )(*xy)
         assert np.allclose(hessian, expected.T)
 
-    @pytest.mark.parametrize("argnum", [None, (0,)])
-    def test_error_wrong_diagonal_shifts(self, argnum):
-        """Test that an error is raised if the number of diagonal shifts does
-        not match the required number (`len(trainable_params)` or `len(argnum)`)."""
-
-        with qml.queuing.AnnotatedQueue() as q:
-            qml.RX(0.4, wires=0)
-            qml.CRY(0.9, wires=[0, 1])
-
-        tape = qml.tape.QuantumScript.from_queue(q)
-        with pytest.raises(ValueError, match="sets of shift values for diagonal entries"):
-            qml.gradients.param_shift_hessian(tape, argnum=argnum, diagonal_shifts=[])
-
-    @pytest.mark.parametrize("argnum", [None, (0, 1)])
-    def test_error_wrong_offdiagonal_shifts(self, argnum):
-        """Test that an error is raised if the number of offdiagonal shifts does
-        not match the required number (`len(trainable_params)` or `len(argnum)`)."""
-
-        with qml.queuing.AnnotatedQueue() as q:
-            qml.RX(0.4, wires=0)
-            qml.CRY(0.9, wires=[0, 1])
-            qml.RX(-0.4, wires=0)
-
-        tape = qml.tape.QuantumScript.from_queue(q)
-        with pytest.raises(ValueError, match="sets of shift values for off-diagonal entries"):
-            qml.gradients.param_shift_hessian(tape, argnum=argnum, off_diagonal_shifts=[])
-
 
 class TestInterfaces:
     """Test the param_shift_hessian method on different interfaces"""
 
+    @pytest.mark.skip("Requires Torch integration for new return types")
     @pytest.mark.torch
     def test_hessian_transform_with_torch(self):
         """Test that the Hessian transform can be used with Torch (1d -> 1d)"""
@@ -1283,6 +1762,7 @@ class TestInterfaces:
 
         assert np.allclose(expected, hess.detach())
 
+    @pytest.mark.skip("Requires Torch integration for new return types")
     @pytest.mark.torch
     def test_hessian_transform_is_differentiable_torch(self):
         """Test that the 3rd derivate can be calculated via auto-differentiation in Torch
@@ -1330,7 +1810,7 @@ class TestInterfaces:
         circuit.interface = "jax"
         hess = qml.gradients.param_shift_hessian(circuit, argnums=[0])(x_jax)
 
-        assert np.allclose(expected, hess)
+        assert np.allclose(qml.math.transpose(expected, (1, 2, 0)), hess)
 
     @pytest.mark.jax
     @pytest.mark.slow
@@ -1352,10 +1832,16 @@ class TestInterfaces:
         x_jax = jax.numpy.array([0.1, 0.2])
 
         expected = qml.jacobian(qml.jacobian(qml.jacobian(circuit)))(x)
-        circuit.interface = "jax"
-        jax_deriv = jax.jacobian(qml.gradients.param_shift_hessian(circuit))(x_jax)
 
-        assert np.allclose(expected, jax_deriv)
+        def cost_fn(x):
+            hess = qml.gradients.param_shift_hessian(circuit)(x)
+            hess = qml.math.stack([qml.math.stack(row) for row in hess])
+            return hess
+
+        circuit.interface = "jax"
+        jax_deriv = jax.jacobian(cost_fn)(x_jax)
+
+        assert np.allclose(qml.math.transpose(expected, (1, 2, 0, 3)), jax_deriv)
 
     @pytest.mark.tf
     @pytest.mark.slow
@@ -1365,7 +1851,7 @@ class TestInterfaces:
 
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="parameter-shift", max_diff=2)
+        @qml.qnode(dev, max_diff=2)
         def circuit(x):
             qml.RX(x[1], wires=0)
             qml.RY(x[0], wires=0)
@@ -1378,9 +1864,9 @@ class TestInterfaces:
         expected = qml.jacobian(qml.jacobian(circuit))(x_np)
         circuit.interface = "tf"
         with tf.GradientTape():
-            hess = qml.gradients.param_shift_hessian(circuit)(x_tf)[0]
+            hess = qml.gradients.param_shift_hessian(circuit)(x_tf)
 
-        assert np.allclose(expected, hess)
+        assert np.allclose(qml.math.transpose(expected, (1, 2, 0)), hess)
 
     @pytest.mark.tf
     @pytest.mark.slow
@@ -1391,7 +1877,7 @@ class TestInterfaces:
 
         dev = qml.device("default.qubit", wires=2)
 
-        @qml.qnode(dev, diff_method="parameter-shift", max_diff=3)
+        @qml.qnode(dev, max_diff=3)
         def circuit(x):
             qml.RX(x[1], wires=0)
             qml.RY(x[0], wires=0)
@@ -1405,6 +1891,8 @@ class TestInterfaces:
         circuit.interface = "tf"
         with tf.GradientTape() as tf_tape:
             hessian = qml.gradients.param_shift_hessian(circuit)(x_tf)[0]
+            hessian = qml.math.stack([qml.math.stack(row) for row in hessian])
+
         tensorflow_deriv = tf_tape.jacobian(hessian, x_tf)
 
-        assert np.allclose(expected, tensorflow_deriv)
+        assert np.allclose(qml.math.transpose(expected, (1, 2, 0, 3)), tensorflow_deriv)
